requirements.txt
python-dateutil
rich
plyer
fastapi
uvicorn
